requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
open3d==0.17.0
Pillow==10.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0 